Now, when registering, the data gets into the table: user_base
