+++ client/pom.xml
@@ -20,8 +20,9 @@
         <dependency>
             <groupId>org.xerial</groupId>
             <artifactId>sqlite-jdbc</artifactId>
-            <version>3.36.0.3</version>
+            <version>3.34.0</version>
         </dependency>
+
 
     </dependencies>
 </project>
