+++ client/src/main/java/client/Controller.java
@@ -174,6 +174,7 @@ public class Controller implements Initializable{
         if (socket == null || socket.isClosed()) {
             connect();
         }
+
         try {
             out.writeUTF(msg);
         } catch (IOException e) {
